Changed pathing issues with CSS to make files work on local
Changed the external link to the CSS within the HTML file to properly
path to the CSS file and have it display properly on my desktop.

For reference, href="/css/main.css" with a slash to start will pull the
path back to the root folder, which the css folder is not located it.
Since the css folder is in the same folder as the HTML file,
href="css/main.css" works.